Merge pull request #7 from Hengjn08/databaseUpdate

+++ app/src/main/java/com/example/unicanteen/database/OrderList.kt
@@ -5,7 +5,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import java.sql.Date
+import java.util.Date
 
 @Entity(
     tableName = "orderList",
@@ -13,17 +13,21 @@ import java.sql.Date
         ForeignKey(entity = Seller::class, parentColumns = ["sellerId"], childColumns = ["sellerId"], onDelete = ForeignKey.CASCADE),
         ForeignKey(entity = FoodList::class, parentColumns = ["foodId"], childColumns = ["foodId"], onDelete = ForeignKey.CASCADE),
         ForeignKey(entity = User::class, parentColumns = ["userId"], childColumns = ["userId"], onDelete = ForeignKey.CASCADE),
-        ForeignKey(entity = Order::class, parentColumns = ["orderId"], childColumns = ["orderId"], onDelete = ForeignKey.CASCADE) // Link to Order entity
-
+        ForeignKey(entity = Order::class, parentColumns = ["orderId"], childColumns = ["orderId"], onDelete = ForeignKey.CASCADE)
     ],
-    indices = [Index(value = ["sellerId"]), Index(value = ["foodId"]), Index(value = ["userId"]), Index(value = ["orderId"])]
+    indices = [
+        Index(value = ["sellerId"]),
+        Index(value = ["foodId"]),
+        Index(value = ["userId"]),
+        Index(value = ["orderId"])
+    ]
 )
 data class OrderList(
-    @PrimaryKey(autoGenerate = true) val orderListId: Long = 0, // Primary key
-    @ColumnInfo(name = "sellerId") val sellerId: Long, // Foreign key to Seller entity
-    @ColumnInfo(name = "foodId") val foodId: Long, // Foreign key to FoodList entity
-    @ColumnInfo(name = "userId") val userId: Long, // Foreign key to User entity
-    @ColumnInfo(name = "orderId") val orderId: Long, // Foreign key to Order entity
+    @PrimaryKey(autoGenerate = true) val orderListId: Int = 0, // Keep as Int for consistency with the database
+    @ColumnInfo(name = "sellerId") val sellerId: Int, // Foreign key to Seller entity
+    @ColumnInfo(name = "foodId") val foodId: Int, // Foreign key to FoodList entity
+    @ColumnInfo(name = "userId") val userId: Int, // Foreign key to User entity
+    @ColumnInfo(name = "orderId") val orderId: Int, // Foreign key to Order entity
 
     @ColumnInfo(name = "qty") val qty: Int, // Quantity of the food item ordered
     @ColumnInfo(name = "totalPrice") val totalPrice: Double, // Total price for this order item
