+++ app/src/main/java/com/example/unicanteen/database/Payment.kt
@@ -5,7 +5,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import java.sql.Date
+import java.util.Date
 
 
 @Entity(
@@ -17,9 +17,9 @@ import java.sql.Date
     indices = [Index(value = ["orderId"]), Index(value = ["userId"])]
 )
 data class Payment(
-    @PrimaryKey(autoGenerate = true) val paymentId: Long = 0, // Primary key
-    @ColumnInfo(name = "orderId") val orderId: Long, // Foreign key to Order entity
-    @ColumnInfo(name = "userId") val userId: Long, // Foreign key to User entity
+    @PrimaryKey(autoGenerate = true) val paymentId: Int = 0, // Primary key
+    @ColumnInfo(name = "orderId") val orderId: Int, // Foreign key to Order entity
+    @ColumnInfo(name = "userId") val userId: Int, // Foreign key to User entity
 
     @ColumnInfo(name = "totalAmt") val totalAmt: Double, // Total amount for the payment
     @ColumnInfo(name = "createDate") val createDate: Date, // Date the payment was created
