+++ app/src/main/java/com/example/unicanteen/database/PaymentDetails.kt
@@ -5,7 +5,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import java.sql.Date
+import java.util.Date
 
 @Entity(
     tableName = "paymentDetails",
@@ -18,11 +18,11 @@ import java.sql.Date
     indices = [Index(value = ["orderId"]), Index(value = ["orderListId"]), Index(value = ["sellerId"]), Index(value = ["paymentId"])]
 )
 data class PaymentDetails(
-    @PrimaryKey(autoGenerate = true) val paymentDetailsId: Long = 0, // Primary key
-    @ColumnInfo(name = "orderId") val orderId: Long, // Foreign key to Order entity
-    @ColumnInfo(name = "orderListId") val orderListId: Long, // Foreign key to OrderList entity
-    @ColumnInfo(name = "sellerId") val sellerId: Long, // Foreign key to Seller entity
-    @ColumnInfo(name = "paymentId") val paymentId: Long, // Foreign key to Payment entity
+    @PrimaryKey(autoGenerate = true) val paymentDetailsId: Int = 0, // Primary key
+    @ColumnInfo(name = "orderId") val orderId: Int, // Foreign key to Order entity
+    @ColumnInfo(name = "orderListId") val orderListId: Int, // Foreign key to OrderList entity
+    @ColumnInfo(name = "sellerId") val sellerId: Int, // Foreign key to Seller entity
+    @ColumnInfo(name = "paymentId") val paymentId: Int, // Foreign key to Payment entity
 
     @ColumnInfo(name = "amount") val amount: Double, // Amount paid to the seller
     @ColumnInfo(name = "createDate") val createDate: Date, // Date the payment detail was created
